Make sure Bower components go where they are expected
The test index.html assumes bower components will be in `./bower_components`.  If other generators include .bowerrc files (e.g. generator-webapp), this may not end up being the Bower components directory.  An alternative to this would be to use Bower's own logic to determine the path to the components directory before processing the index.html template.